couple-app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==1.4.49
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic[email]
alembic
bcrypt==3.2.0
pytest
httpx
pytest
python-multipart
requests
pydantic_settings
google-auth
google-auth-oauthlib
google-cloud-storage


